Self Problem 01 Calculate Simple Interest

+++ AreaOfCircle.java
@@ -7,7 +7,7 @@ import java.util.*;
       public static void main(String[] ars){
         Scanner s = new Scanner(System.in);
         System.out.println("Enter the value of Radius of circle: ");
-        double r = s.nextInt();
+        double r = s.nextDouble();
         double area = areaOfCircle(r);
         System.out.println(area);
       }
